% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{cov.inner.shift}
\alias{cov.inner.shift}
\title{Estimate the covariance between the pi(l1,k1) entry and pi(l2,k2) entry
when the covariance for the targeted distribution is unknown}
\usage{
cov.inner.shift(
  Var.vec,
  l1,
  k1,
  l2,
  k2,
  X.l1,
  X.k1,
  X.l2,
  X.k2,
  Pred.mat.target,
  Proj.array
)
}
\arguments{
\item{Var.vec}{Variance of residuals in groups, of length \eqn{L}}

\item{l1}{Index l1}

\item{k1}{Index k1}

\item{l2}{Index l2}

\item{k2}{Index k2}

\item{X.l1}{Design matrix of group \eqn{l1} in training data}

\item{X.k1}{Design matrix of group \eqn{k1} in training data}

\item{X.l2}{Design matrix of group \eqn{l2} in training data}

\item{X.k2}{Design matrix of group \eqn{k2} in training data}

\item{Pred.mat.target}{Predicted outcome matrix for target design matrix
using L fitted coefficients, of dimension \eqn{n.target} x \eqn{L}}

\item{Proj.array}{Projection directions, of dimension \eqn{L} x \eqn{L} x \eqn{p}}
}
\value{
covariance between the pi(l1, k1) entry and pi(l2, k2) entry
}
\description{
Estimate the covariance between the pi(l1,k1) entry and pi(l2,k2) entry
when the covariance for the targeted distribution is unknown
}
