% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{decide_delta}
\alias{decide_delta}
\title{Decide ridge penalty data-dependently}
\usage{
decide_delta(Gamma, step_delta = 0.1, MAX_iter = 100, verbose = FALSE)
}
\arguments{
\item{Gamma}{Weight Matrix}

\item{step_delta}{The step size of searching delta (default = 0.1)}

\item{MAX_iter}{Maximum of iterations for searching}

\item{verbose}{Print information about delta and reward (default = \code{FALSE})}
}
\value{
\item{delta}{The data-dependent ridge penalty}
}
\description{
Decide ridge penalty data-dependently
}
