% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{Gamma.shift}
\alias{Gamma.shift}
\title{Bias correction for initial estimator of Gamma when covariate shifts}
\usage{
Gamma.shift(plug.in, X.l, X.k, omega.l, omega.k, Y.l, Y.k, Pred.l, Pred.k)
}
\arguments{
\item{plug.in}{Initial estimator of Gamma target, of dimension \eqn{L} x \eqn{L}}

\item{X.l}{Design matrix of label \eqn{l} in training data, of dimension \eqn{n_l} x \eqn{p}}

\item{X.k}{Design matrix of label \eqn{k} in training data, of dimension \eqn{n_k} x \eqn{p}}

\item{omega.l}{The l-th column of Omega matrix}

\item{omega.k}{The k-th column of Omega matrix}

\item{Y.l}{Outcome vector of label \eqn{l} in training data, of length \eqn{n_l}}

\item{Y.k}{Outcome vector of label \eqn{k} in training data, of length \eqn{n_k}}

\item{Pred.l}{Predicted outcome vector of label \eqn{l} in training data}

\item{Pred.k}{Predicted outcome vector of label \eqn{k} in training data}
}
\value{
\item{est}{The proposed bias-corrected estimator of Gamma target}
\item{proj.lk}{The projection direction of index (l, k)}
\item{proj.kl}{The projection direction of index (k, l)}
}
\description{
Bias correction for initial estimator of Gamma when covariate shifts
}
