% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{infer.Maximin}
\alias{infer.Maximin}
\title{Inference method for Maximin}
\usage{
\method{infer}{Maximin}(object, gen.size = 500, delta = -1, threshold = 2, alpha = 0.01, ...)
}
\arguments{
\item{object}{Object of class inheriting from "Maximin"}

\item{gen.size}{The generated sample size}

\item{delta}{The ridge penalty. If set as negative value, the penalty is decided data-dependently. (Default = -1)}

\item{threshold}{Should generated samples be filter or not? If 0, do not filter;
if 1, use chi-square threshold to filter;
if 2, use normal threshold to filter. (Default = 2)}

\item{alpha}{confidence value to select generated samples}

\item{...}{further arguments passed}
}
\value{
\item{delta}{The ridge penalty used}
\item{weight}{The weight vector for groups, of length \eqn{L}}
\item{point}{The point estimator of the linear contrast}
\item{mm.effect}{The aggregated maximin effect (coefficients), of length \eqn{p} or \eqn{p+1}}
\item{CI}{Confidence interval for the linear contrast}
}
\description{
Point estimator and Confidence interval based on Maximin object
}
\examples{
\donttest{
## number of groups
L=2
## dimension
p=500

## mean vector for source
mean.source = rep(0, p)
## covariance matrix for source
A1gen <- function(rho,p){
  A1=matrix(0,p,p)
  for(i in 1:p){
    for(j in 1:p){
      A1[i,j]<-rho^(abs(i-j))
    }
  }
  return(A1)
}
cov.source = A1gen(0.6, p)

## 1st group's source data
n1 = 500
X1 = MASS::mvrnorm(n1, mu=mean.source, Sigma=cov.source)
b1 = rep(0, p)
b1[1:10] = seq(1:10)/40 # true coef for 1st group
Y1 = X1\%*\%b1 + rnorm(n1)

## 2nd group's source data
n2 = 400
X2 = MASS::mvrnorm(n2, mu=mean.source, Sigma=cov.source)
b2 = rep(0, p)
b2[1:10] = -seq(1:10)/40 # true coef for 2nd group
Y2 = X2\%*\%b2 + rnorm(n2)

## Target Data, covariate shift
n.target = 500
mean.target = rep(0, p)
cov.target = cov.source
for(i in 1:p) cov.target[i, i] = cov.target[i, i] + 0.1
for(i in 1:5){
  for(j in 1:5){
    if(i!=j) cov.target[i, j] = 0.9
  }
}
X.target = MASS::mvrnorm(n.target, mu=mean.target, Sigma=cov.target)

## loading
loading = rep(0, p)
loading[1:5] = 1

## call
mm <- Maximin(list(X1, X2), list(Y1, Y2), X.target, loading, covariate.shift = TRUE)
mmInfer <- infer(mm, gen.size=100, delta=0)
}
}
