% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{MaximinInfer}
\alias{MaximinInfer}
\title{Maximin Inference}
\usage{
MaximinInfer(
  Xlist,
  Ylist,
  loading,
  X.target = NULL,
  cov.target = NULL,
  covariate.shift = TRUE,
  lam.value = c("CV", "CV.min"),
  intercept = TRUE,
  intercept.loading = FALSE,
  gen.size = 500,
  delta = 0,
  threshold = c(0, 1, 2),
  alpha = 0.05,
  alpha.thres = 0.01
)
}
\arguments{
\item{Xlist}{list of design matrix for source data, of length \eqn{L}}

\item{Ylist}{list of outcome vector for source data, of length \eqn{L}}

\item{loading}{Loading, of length \eqn{p}}

\item{X.target}{Design matrix for target data, of dimension \eqn{n.target} x \eqn{p} (default = `NULL`)}

\item{cov.target}{Covariance matrix for target data, of dimension \eqn{p} x \eqn{p} (default = `NULL`)}

\item{covariate.shift}{Covariate shifts or not between source and target data (default = `TRUE`)}

\item{lam.value}{The method to be used to obtain Lasso estimator of high-dimensional regression vector for each group}

\item{intercept}{Should intercept be fitted for the initial estimator (default = `TRUE`)}

\item{intercept.loading}{Should intercept be included for the loading (default = `FALSE`)}

\item{gen.size}{The generating sample size (Default = 500)}

\item{delta}{The ridge penalty (Default = 0)}

\item{threshold}{Should generated samples be filtered or not?
If 0, use normal threshold to filter;
if 1, use chi-square threshold to filter;
if 2, do not filter. (Default = 0)}

\item{alpha}{confidence value to construct confidence interval (Default = 0.05)}

\item{alpha.thres}{confidence value to select generated samples (Default = 0.01)}
}
\value{
\item{weight}{The weight vector for groups, of length \eqn{L}}
\item{point}{The point estimator of the linear combination}
\item{mm.effect}{The aggregated maximin effect (coefficients), of length \eqn{p} or \eqn{p+1}}
\item{CI}{Confidence interval for the linear combination}
}
\description{
`MaximinInfer` is a wrapper for class `Maximin` and the method `infer`.
}
\details{
The algorithm implemented scenarios with or without covariate shift. If `cov.target` is specified,
the `X.target` will be ignored; if not, while `X.target` is specified, `cov.target` will be estimated
by `X.target`. If both are not specified, the algorithm will automatically set `covariate.shift` as
`FALSE`.
}
\examples{
\donttest{
## number of groups
L=2
## dimension
p=500

## mean vector for source
mean.source = rep(0, p)
## covariance matrix for source
A1gen <- function(rho,p){
  A1=matrix(0,p,p)
  for(i in 1:p){
    for(j in 1:p){
      A1[i,j]<-rho^(abs(i-j))
    }
  }
  return(A1)
}
cov.source = A1gen(0.6, p)

## 1st group's source data
n1 = 500
X1 = MASS::mvrnorm(n1, mu=mean.source, Sigma=cov.source)
b1 = rep(0, p)
b1[1:10] = seq(1:10)/40 # true coef for 1st group
Y1 = X1\%*\%b1 + rnorm(n1)

## 2nd group's source data
n2 = 400
X2 = MASS::mvrnorm(n2, mu=mean.source, Sigma=cov.source)
b2 = rep(0, p)
b2[1:10] = -seq(1:10)/40 # true coef for 2nd group
Y2 = X2\%*\%b2 + rnorm(n2)

## Target Data, covariate shift
n.target = 500
mean.target = rep(0, p)
cov.target = cov.source
for(i in 1:p) cov.target[i, i] = cov.target[i, i] + 0.1
for(i in 1:5){
  for(j in 1:5){
    if(i!=j) cov.target[i, j] = 0.9
  }
}
X.target = MASS::mvrnorm(n.target, mu=mean.target, Sigma=cov.target)

## loading
loading = rep(0, p)
loading[1:5] = 1

## call
out <- MaximinInfer(list(X1, X2), list(Y1, Y2), loading, X.target, covariate.shift = TRUE)
out$CI
}
}
