% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{measure_instability}
\alias{measure_instability}
\title{instability measurement}
\usage{
measure_instability(
  object,
  delta = 0,
  gen.size = 500,
  threshold = c(0, 1, 2),
  alpha.thres = 0.01
)
}
\arguments{
\item{object}{Object of class inheriting from "Maximin"}

\item{delta}{The ridge penalty (Default = 0)}

\item{gen.size}{The generating sample size (Default = 500)}

\item{threshold}{Should generated samples be filtered or not?
If 0, use normal threshold to filter;
if 1, use chi-square threshold to filter;
if 2, do not filter. (Default = 0)}

\item{alpha.thres}{confidence value to select generated samples (Default = 0.01)}
}
\value{
\item{measure}{The measurement of instability}
}
\description{
compute the instability measurement given a specific ridge penalty
}
