% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{opt.weight}
\alias{opt.weight}
\title{Compute Ridge-type weight vector}
\usage{
opt.weight(Gamma, delta, report.reward = TRUE)
}
\arguments{
\item{Gamma}{regression covariance matrix, of dimension \eqn{L} x \eqn{L}.}

\item{delta}{the ridge penalty level, non-positive.}

\item{report.reward}{the reward is computed or not (Default = \code{TRUE})}
}
\value{
\item{weight}{the minimizer \eqn{\gamma}}
\item{reward}{the value of penalized reward}
}
\description{
\eqn{\argmin_{\gamma \in \Delta} \gamma^T (\Gamma + \delta * I) \gamma}.
}
