% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data for Analysis}
\format{
list with source data and target data, which are:
\describe{
\item{X1}{Design matrix for the 1st group source data}
\item{X2}{Design matrix for the 2nd group source data}
\item{Y1}{Outcome vector for the 1st group source data}
\item{Y2}{Outcome vector for the 2nd group source data}
\item{X.target}{Design matrix for the target data}
}
}
\usage{
sample_data
}
\description{
Datasets for the simple testing and running examples. The data
is heterogenous with 2 groups, and covariates shift between target data
and source data.
}
\keyword{datasets}
