% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{decide_delta}
\alias{decide_delta}
\title{decide delta}
\usage{
decide_delta(object, step_delta = 0.1, MAX_iter = 100, verbose = FALSE)
}
\arguments{
\item{object}{Object of class inheriting from "Maximin"}

\item{step_delta}{The step size of searching delta (Default = 0.1)}

\item{MAX_iter}{Maximum of iterations for searching (Default = 100)}

\item{verbose}{Print information about delta and reward (Default = `FALSE`)}
}
\value{
\item{delta}{The data-dependent ridge penalty}
\item{reward.ratio}{The ratio of penalized reward over non-penalized reward}
}
\description{
`decide_delta` will tell if the estimator is stable or not without ridge penalty
at first. If instable, it picks a ridge penalty data-dependently.
}
