% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maximin.R
\name{decide_delta}
\alias{decide_delta}
\title{Decide ridge penalty data-dependently}
\usage{
decide_delta(
  obj,
  gen.size = 500,
  step_delta = 0.1,
  MAX_iter = 100,
  verbose = FALSE
)
}
\arguments{
\item{obj}{The returned list of Maximin}

\item{gen.size}{The generating sample size (Default = 500)}

\item{step_delta}{The step size of searching delta (Default = 0.1)}

\item{MAX_iter}{Maximum of iterations for searching (Default = 100)}

\item{verbose}{Print information about delta and reward (Default = \code{FALSE})}
}
\value{
\item{delta}{The data-dependent ridge penalty}
\item{reward.ratio}{The ratio of penalized reward over non-penalized reward}
}
\description{
To tell if the estimator is stable or not without ridge penalty
at first. If instable, it picks a ridge penalty data-dependently.
}
