% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{infer}
\alias{infer}
\title{Maximin Inference}
\usage{
infer(object, ...)
}
\arguments{
\item{object}{a "Maximin" object}

\item{...}{additional arguments affecting inference}
}
\description{
`infer` is a generic function for inference for Maximin model.
}
