% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maximin.R
\name{Maximin}
\alias{Maximin}
\title{Class Maximin}
\usage{
Maximin(
  Xlist,
  Ylist,
  X.target = NULL,
  loading,
  cov.target = NULL,
  covariate.shift = TRUE,
  lam.value = c("CV", "CV.min", "scalreg", "slim"),
  intercept = TRUE,
  intercept.loading = FALSE
)
}
\arguments{
\item{Xlist}{list of design matrix for source data, of length \eqn{L}.}

\item{Ylist}{list of outcome vector for source data, of length \eqn{L}}

\item{X.target}{Design matrix for target data, of dimension \eqn{n.target} x \eqn{p} (default = `NULL`)}

\item{loading}{Loading, of length \eqn{p}}

\item{cov.target}{Covariance matrix for target data, of dimension \eqn{p} x \eqn{p}. If set as `NULL`, `cov.target` is unknown. (default = `NULL`)}

\item{covariate.shift}{Covariate shifts or not between source and target data (default = `TRUE``)}

\item{lam.value}{The method to be used to obtain each group's intial estimator}

\item{intercept}{Should intercept be fitted for the initial estimator (default = `TRUE`)}

\item{intercept.loading}{Should intercept be included for the loading (default = `FALSE`)}
}
\value{
`Maximin` returns an object of class "Maximin". The function `infer` is used to do further inference.\\
An object of class "Maximin" is a list containing the following components.
\item{Gamma.prop}{The proposed debiased Weight matrix}
\item{Coef.est}{The initial estimators for each group}
\item{Point.vec}{The point estimator for each group}
\item{L}{The number of groups}
\item{gen.mu}{The mean vector for sampling method}
\item{gen.Cov}{The variance matrix for sampling method}
}
\description{
Class Maximin
}
