% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maximin.R
\name{measure_instability}
\alias{measure_instability}
\title{measurement of instability}
\usage{
measure_instability(
  obj,
  delta = 0,
  gen.size = 500,
  threshold = 0,
  alpha.thres = 0.01
)
}
\arguments{
\item{obj}{The returned list of Maximin}

\item{delta}{The ridge penalty (Default = 0)}

\item{gen.size}{The generating sample size (Default = 500)}

\item{threshold}{Should generated samples be filtered or not?
if 0, use normal threshold to filter;
if 1, use chi-square threshold to filter;
if 2, do not filter. (Default = 0)}

\item{alpha.thres}{The confidence value to select generated samples (Default = 0.01)}
}
\value{
The measurement of instability
}
\description{
compute the instability measurement given a specific ridge penalty
}
